+++ src/games/index.ts
@@ -19,6 +19,7 @@ import { AccastaGame, IAccastaState } from "./accasta";
 import { EpamGame, IEpamState } from "./epam";
 import { TaijiGame, ITaijiState } from "./taiji";
 import { BreakthroughGame, IBreakthroughState } from "./breakthrough";
+import { FabrikGame, IFabrikState } from "./fabrik";
 
 export {
     APGamesInformation, GameBase, IAPGameState,
@@ -40,7 +41,8 @@ export {
     AccastaGame, IAccastaState,
     EpamGame, IEpamState,
     TaijiGame, ITaijiState,
-    BreakthroughGame, IBreakthroughState
+    BreakthroughGame, IBreakthroughState,
+    FabrikGame, IFabrikState,
 };
 
 const games = new Map<string, typeof AmazonsGame | typeof BlamGame | typeof CannonGame |
@@ -48,11 +50,12 @@ const games = new Map<string, typeof AmazonsGame | typeof BlamGame | typeof Cann
                               typeof VolcanoGame | typeof MvolcanoGame | typeof ChaseGame |
                               typeof AbandeGame | typeof CephalopodGame | typeof LinesOfActionGame |
                               typeof PikemenGame | typeof OrdoGame | typeof AttangleGame |
-                              typeof AccastaGame | typeof EpamGame | typeof TaijiGame | typeof BreakthroughGame>();
+                              typeof AccastaGame | typeof EpamGame | typeof TaijiGame | typeof BreakthroughGame |
+                              typeof FabrikGame>();
 // Manually add each game to the following array
 [AmazonsGame, BlamGame, CannonGame, MchessGame, HomeworldsGame, EntropyGame, VolcanoGame, MvolcanoGame,
     ChaseGame, AbandeGame, CephalopodGame, LinesOfActionGame, PikemenGame, OrdoGame, AttangleGame,
-    AccastaGame, EpamGame, TaijiGame, BreakthroughGame].forEach((g) => {
+    AccastaGame, EpamGame, TaijiGame, BreakthroughGame, FabrikGame].forEach((g) => {
     if (games.has(g.gameinfo.uid)) {
         throw new Error("Another game with the UID '" + g.gameinfo.uid + "' has already been used. Duplicates are not allowed.");
     }
@@ -100,6 +103,8 @@ export function GameFactory(game: string, ...args: any[]): GameBase|undefined {
             return new TaijiGame(...args);
         case "breakthrough":
             return new BreakthroughGame(...args);
+        case "fabrik":
+            return new FabrikGame(...args);
     }
     return;
 }